Update c-ares download URL in Dockerfile for proper installation
This PR fixes an issue with the outdated c-ares download link in the following file:
TargetMicroservices/socialNetwork/Dockerfile.

The previous download link for c-ares-1.25.0.tar.gz was resulting in a 404 Not Found error. The URL has been updated to the correct location from the official c-ares GitHub release:

Old URL:
https://c-ares.org/download/c-ares-1.25.0.tar.gz

New URL:
https://github.com/c-ares/c-ares/releases/download/cares-1_25_0/c-ares-1.25.0.tar.gz

Updating the link ensures that the Docker build process completes successfully and that the required c-ares dependency is installed without issues.